refactor markdown viewer
