+++ src/app/page.tsx
@@ -9,11 +9,11 @@ export default function Home() {
   const { selectedModel } = useControlContext();
 
   return (
-    <main className="grid md:grid-cols-12 max-h-screen h-screen max-w-6xl m-auto gap-4 p-2 md:p-4 my-12 md:my-0">
+    <main className="grid md:grid-cols-12 max-h-screen md:h-screen max-w-6xl m-auto gap-4 p-2 md:p-4 my-12 md:my-0">
       <div className="relative md:col-span-3 h-[95vh]">
         <ControlContainer />
       </div>
-      <div className="md:col-span-9 h-[95vh]">
+      <div className="md:col-span-9">
         {selectedModel === "gemini-pro" ? (
           <ChatContainer />
         ) : (
